+++ main.cu
@@ -5,7 +5,7 @@ int main(){
 	cudaSetDevice(1);
 	cudaDeviceSynchronize();
 	CFD cfd = CFD(1000, 10, 10);
-	double maxTime = 10;
+	float maxTime = 10;
 	cfd.step(maxTime);
 	return 0;
 }
